refactor: relocate common ui components
