+++ src/components/common/NavigationBar.tsx
